+++ NEWS.md
@@ -1,5 +1,8 @@
+# v1.3.2 发布(2018/11/28)
+Add class RecyclerViewStickyHeadItemDecoration，auto make group view type into sticky head mode。
+
 # v1.3.1 发布(2018/08/27)
-fix ItemDecoration Grid mode horizontal offset bug.
+Fix ItemDecoration Grid mode horizontal offset bug.
 
 # v1.3 发布(2018/04/13)
 Combine [RecyclerViewDecoration](https://github.com/arjinmc/RecyclerViewDecoration) into this lib，expand RecyclerViewGroupItemDecoration，Release this lib to the public。